refactor(pipelines): remove implementations submodule and sarscov2 pipeline
Created un-needed nesting.
