--- src/cherami/pipelines/implementations/__init__.py
@@ -1,9 +1,0 @@
-from cherami.pipelines.implementations.amr import AmrPipeline
-from cherami.pipelines.implementations.orange_box import OrangeBoxPipeline
-from cherami.pipelines.implementations.sarscov2 import SARSCoV2Pipeline
-
-__all__ = [
-    "AmrPipeline",
-    "OrangeBoxPipeline",
-    "SARSCoV2Pipeline",
-]
